requirements.txt
absl-py==1.4.0
alabaster==0.7.13
antlr4-python3-runtime==4.9.3
astunparse==1.6.3
attrs==23.1.0
Babel==2.12.1
boto3==1.26.148
botocore==1.29.148
build==0.10.0
cachetools==5.3.1
catkin==0.8.10
certifi==2022.12.7
chardet==4.0.0
charset-normalizer==3.1.0
click==8.1.3
coloredlogs==15.0.1
contourpy==1.0.7
coverage==5.3.1
cycler==0.10.0
Deprecated==1.2.14
docutils==0.17.1
einops==0.3.2
flatbuffers==23.5.26
fonttools==4.39.4
future==0.18.3
gast==0.4.0
gencpp==0.7.0
geneus==3.0.0
genlisp==0.4.18
genmsg==0.6.0
gennodejs==2.0.1
genpy==0.6.15
google-auth==2.19.1
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.54.2
h5py==3.8.0
humanfriendly==10.0
hydra-core==1.3.2
idna==2.10
imageio==2.31.0
imagesize==1.4.1
imutils==0.5.4
jax==0.4.11
Jinja2==3.1.2
jmespath==1.0.1
json-tricks==3.16.1
jsonschema==4.17.3
keras==2.12.0
kiwisolver==1.4.4
libclang==16.0.0
Markdown==3.4.3
markdown-it-py==2.2.0
MarkupSafe==2.1.3
matplotlib==3.7.1
mdurl==0.1.2
ml-dtypes==0.2.0
mpmath==1.3.0
numpy==1.23.0
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
oauthlib==3.2.2
omegaconf==2.3.0
onnx==1.13.0
onnx-simplifier==0.4.31
onnxruntime==1.13.1
opencv-python==4.7.0.72
opt-einsum==3.3.0
packaging==23.1
Pillow==9.5.0
pip-tools==6.13.0
pipdeptree==2.9.0
protobuf==3.20.3
psutil==5.9.5
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycocotools==2.0.4
pyDeprecate==0.3.2
Pygments==2.15.1
pyparsing==2.4.5
pyproject_hooks==1.0.0
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==1.0.0
PyYAML==6.0
rapidfuzz==3.1.1
requests==2.31.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rich==13.4.1
roboflow==1.0.9
rsa==4.9
s3transfer==0.6.1
scipy==1.10.1
six==1.16.0
snowballstemmer==2.2.0
Sphinx==4.0.3
sphinx-rtd-theme==1.2.1
sphinxcontrib-applehelp==1.0.4
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.1
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
stringcase==1.2.0
super-gradients==3.1.0
supervision==0.8.0
sympy==1.12
tensorboard==2.12.3
tensorboard-data-server==0.7.0
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-io-gcs-filesystem==0.32.0
termcolor==1.1.0
tomli==2.0.1
torch==1.13.1
torchmetrics==0.8.0
torchvision==0.14.1
tqdm==4.65.0
treelib==1.6.1
typing_extensions==4.6.3
urllib3==1.26.16
Werkzeug==2.3.4
wget==3.2
wrapt==1.14.1
